YAN-54 Исправлены критические замечания

+++ packages/server/services/forum/reaction.ts
@@ -3,6 +3,7 @@ import { Sequelize } from 'sequelize-typescript'
 import { xssErrorHandler, xssValidator } from '../../middlewares/xssValidation'
 import { Reaction } from '../../models/forum/reaction'
 import { checkAuth } from '../../middlewares/checkAuth'
+import { sequelize } from '../../db'
 
 export const forumReactionRoute = Router()
   .use(express.json())
@@ -24,37 +25,50 @@ export const forumReactionRoute = Router()
     }
   })
   .post('/', xssValidator(), xssErrorHandler, async (req: Request, res: Response) => {
-    const { unified, topicId, userId } = req.body
+    const { unified, topicId } = req.body
+    const userId = res.locals.user?.id
 
     try {
-      const existingReaction = await Reaction.findOne({
-        where: { unified, topicId, userId },
-      })
+      if (userId) {
+        const reaction = await sequelize.transaction(async transaction => {
+          const existingReaction = await Reaction.findOne({
+            where: { unified, topicId, userId },
+            transaction
+          })
 
-      if (existingReaction) {
-        res.status(400).json({ error: 'Такая запись Reaction уже существует' })
+          if (existingReaction) {
+            return existingReaction
+          } else {
+            return await Reaction.create({ unified, topicId, userId }, { transaction })
+          }
+        })
+
+        res.status(200).json(reaction)
+      } else {
+        res.status(401).json({ error: 'Пользователь не аутентифицирован' })
       }
-
-      const reaction = await Reaction.create({ unified, topicId, userId })
-
-      res.status(200).json(reaction)
     } catch (error: any) {
       res.status(500).json({ error: error?.message })
     }
   })
   .delete('/', xssValidator(), xssErrorHandler, async (req: Request, res: Response) => {
-    const { unified, topicId, userId } = req.body
+    const { unified, topicId } = req.body
+    const userId = res.locals.user?.id
 
     try {
-      const deletedCount = await Reaction.destroy({
-        where: { unified, topicId, userId },
-      })
-
-      if (deletedCount === 0) {
-        res.status(404).json({ error: 'Запись не найдена' })
+      if (userId) {
+        const deletedCount = await Reaction.destroy({
+          where: { unified, topicId, userId },
+        })
+  
+        if (deletedCount === 0) {
+          res.status(404).json({ error: 'Запись не найдена' })
+        } else {
+          res.json({ message: 'Запись успешно удалена' })
+        }
+      } else {
+        res.status(401).json({ error: 'Пользователь не аутентифицирован' })
       }
-
-      res.json({ message: 'Запись успешно удалена' })
     } catch (error: any) {
       res.status(500).json({ error: error?.message })
     }
